Fixed The Issue Regarding Test Fail
Steps To Follow -

delete node_modules and do a install again

It will Install Ethers with a version that supports the function that we need.

Fetched the Deployement Addresses directly from the deploy function to aid us in the address

Some test were failing because we were using signer.address instead of the deployed contract address.

Rest updated the code for the ethers version